fix: use short month date format

+++ frontend/interfaces/dateFormatter.ts
@@ -1,5 +1,5 @@
 const formatDate = (date: string) => {
-    const options = {year: 'numeric', month: 'long', day: 'numeric', hour: 'numeric', minute: 'numeric'};
+    const options = {year: 'numeric', month: 'short', day: 'numeric', hour: 'numeric', minute: 'numeric'};
     return new Date(date).toLocaleDateString("en-US", options);
 };
